requirements.txt
# TODO: Update with exact module version
numpy
yolox
opencv-python
loguru
pytorch
